Cleanup URL utils
