+++ src/components/documentBuilder/preview/elementsPreview/Image.tsx
@@ -26,7 +26,7 @@ import documentTreeStyles from "@/components/documentBuilder/preview/documentTre
 import ImagePlaceholder from "@/components/imagePlaceholder/ImagePlaceholder";
 import { isEmpty } from "lodash";
 import Flaps from "@/components/documentBuilder/preview/flaps/Flaps";
-import { isValidUrl } from "@/utils/urlUtils";
+import { urlMatchesProtocol } from "@/utils/urlUtils";
 
 type ImageProps = PreviewComponentProps & {
   elementType: DocumentElementType;
@@ -46,9 +46,7 @@ const Image: React.FunctionComponent<ImageProps> = ({
   ...restPreviewProps
 }) => {
   // If it's not a valid URL, show a placeholder
-  const renderPlaceholder =
-    typeof props.src !== "string" ||
-    !isValidUrl(props.src, { protocols: ["https:"] });
+  const renderPlaceholder = !urlMatchesProtocol(props.src, ["https:"]);
 
   return (
     <div
